refactor(ui): update import path for WelcomeView component
