+++ webview-ui/src/components/kilocode/welcome/WelcomeView.tsx
